Add option to use local spec for generation

+++ scripts/Generate-LatestOpenAPICodeAndDocs.ps1
@@ -1,10 +1,19 @@
-# Download latest spec for the API
+[CmdletBinding()]
+param(
+  [Parameter()]
+  [switch] $UseLocalSpec = $False
+)
+
 $specFile = "spec-v1-swagger.json"
-if (Test-Path "../$specFile")
+if(!$UseLocalSpec)
 {
-  Remove-Item "../$specFile" -Force
+  # Download latest spec for the API
+  if (Test-Path "../$specFile")
+  {
+    Remove-Item "../$specFile" -Force
+  }
+  Invoke-WebRequest -Uri "https://api.youneedabudget.com/papi/$specFile" -OutFile "../$specFile"
 }
-Invoke-WebRequest -Uri "https://api.youneedabudget.com/papi/$specFile" -OutFile "../$specFile"
 
 # Use Docker to codegen dotnetcore based on the swagger spec
 $projRoot = (Get-Item(Get-Location)).Parent.FullName
